changed limit line quantity text format
